refactor(render): move the pack choice and its settings out of PackChain
PackChain carried, beside the frame, the whole of a load short of it:
which pack pack.txt names, how it is found among the candidates or
reported missing, how its settings are laid out and carried over, the
four numbers the screen moves without a load behind them, and when to
read everything again. That block touched the chain twice, to name the
pack in a message, and worked on static fields of its own that the
frame reads one of, the shadow distance the slider moves. It becomes
PackChoice, with those fields. The frame keeps the chain up now, whether
it has been stopped and whether the chain runs at all, and the choice
reaches those three through activate, takeDown, stop and chainWanted
rather than by writing the fields; the reason a frame stops comes back
the other way.
