+++ common/src/main/java/dev/vitrail/settings/SettingsLayers.java
@@ -94,7 +94,7 @@ public final class SettingsLayers {
 	 * Through the {@code String} constructor and not {@code Files.readAllLines}, which THROWS on a
 	 * byte that is not UTF-8. This file is edited by hand, in whichever editor the player has, and
 	 * that throw does not stop at the line it cannot read: it leaves {@code forced}, reaches
-	 * {@code PackChain.open}, and NO PACK LOADS AT ALL, under a message naming neither the file nor
+	 * {@code PackChoice.open}, and NO PACK LOADS AT ALL, under a message naming neither the file nor
 	 * the encoding. What one stray byte costs instead is one character of one value.
 	 * {@code SettingsFile.lines} reads its own file this way already, for the same reason.
 	 * <p>
